+++ opencog/atoms/CMakeLists.txt
@@ -6,6 +6,7 @@ ADD_SUBDIRECTORY (container)
 ADD_SUBDIRECTORY (flow)
 ADD_SUBDIRECTORY (execution)
 ADD_SUBDIRECTORY (grounded)
+ADD_SUBDIRECTORY (parallel)
 ADD_SUBDIRECTORY (pattern)
 ADD_SUBDIRECTORY (reduct)
 ADD_SUBDIRECTORY (truthvalue)
