+++ opencog/atoms/execution/CMakeLists.txt
@@ -7,8 +7,6 @@ ADD_LIBRARY (execution
 	EvaluationLink.cc
 	ExecutionOutputLink.cc
 	Instantiator.cc
-	ParallelLink.cc
-	ThreadJoinLink.cc
 )
 
 # Without this, parallel make will race and crap up the generated files.
@@ -31,7 +29,5 @@ INSTALL (FILES
 	Force.h
 	GroundedProcedureNode.h
 	Instantiator.h
-	ParallelLink.h
-	ThreadJoinLink.h
 	DESTINATION "include/opencog/atoms/execution"
 )
